+++ packages/experiments/src/components/Button/index.ts
@@ -1,5 +1,8 @@
+// Button
 export * from './Button';
 export * from './Button.types';
+
+// Button variants
 export * from './ButtonVariants.types';
 export * from './ActionButton';
 export * from './BaseButton';