+++ packages/utilities/src/dom/raiseClick.ts
@@ -1,7 +1,19 @@
 /** Raises a click event. */
 export function raiseClick(target: Element): void {
-  const event = new Event('MouseEvents');
-
+  const event = createNewEvent('MouseEvents');
   event.initEvent('click', true, true);
   target.dispatchEvent(event);
+}
+
+function createNewEvent(eventName: string): Event {
+  let event;
+  if (typeof Event === 'function') {
+    // Chrome, Opera, Firefox
+    event = new Event(eventName);
+  } else {
+    // IE
+    event = document.createEvent('Event');
+    event.initEvent(eventName, true, true);
+  }
+  return event;
 }