+++ packages/react-cards/CHANGELOG.md
@@ -1,6 +1,13 @@
 # Change Log - @uifabric/react-cards
 
-This log was last generated on Wed, 19 Jun 2019 12:27:03 GMT and should not be manually modified.
+This log was last generated on Thu, 27 Jun 2019 18:25:04 GMT and should not be manually modified.
+
+## 0.107.2
+Thu, 27 Jun 2019 18:25:04 GMT
+
+### Patches
+
+- Updating snapshots after Button changes.
 
 ## 0.107.1
 Wed, 19 Jun 2019 12:27:03 GMT